Tests/LibWeb/Text/input/DOM/DOMStringMap-setter-with-invalid-name.html

<script src="../include.js"></script>
<script>
    test(() => {
        let threw = false;
        try {
            document.body.dataset["'\uDBF8"] = "foo";
        } catch {
            threw = true;
        }
        println("Setting DOMStringMap with an invalid name key throws? " + threw);
    });
</script>
