Tests/LibWeb/Text/expected/DOM/DOMStringMap-setter-with-invalid-name.txt

Setting DOMStringMap with an invalid name key throws? true
